Move as_pretty_string.js (#45356)
* Convert asPrettyString to TS

* Migrate asPrettyString() to plugins/data
