+++ src/legacy/core_plugins/kibana/common/field_formats/types/string.js
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-import { asPrettyString } from '../../utils/as_pretty_string';
+import { asPrettyString } from '../../../../../../plugins/data/common/utils/as_pretty_string';
 import { shortenDottedString } from '../../utils/shorten_dotted_string';
 
 const TRANSFORM_OPTIONS = [
