+++ src/legacy/ui/field_formats/content_types/html_content_type.ts
@@ -18,10 +18,8 @@
  */
 import { escape, isFunction } from 'lodash';
 import { FieldFormatConvert, IFieldFormat, HtmlConventTypeConvert } from '../types';
-
-// @ts-ignore
-import { asPrettyString } from '../../../core_plugins/kibana/common/utils/as_pretty_string';
 import { getHighlightHtml } from '../../../../plugins/data/common/highlight/highlight_html';
+import { asPrettyString } from '../../../../plugins/data/common/utils/as_pretty_string';
 
 const CONTEXT_TYPE = 'html';
