+++ src/legacy/ui/field_formats/content_types/text_content_type.ts
@@ -19,9 +19,7 @@
 
 import { isFunction } from 'lodash';
 import { IFieldFormat, FieldFormatConvert, TextContextTypeConvert } from '../types';
-
-// @ts-ignore
-import { asPrettyString } from '../../../core_plugins/kibana/common/utils/as_pretty_string';
+import { asPrettyString } from '../../../../plugins/data/common/utils/as_pretty_string';
 
 const CONTEXT_TYPE = 'text';
